add slash support for queue list, add page embed util
