+++ commands/queue.js
@@ -1,12 +1,13 @@
 const Command = require("../structures/command.js");
-const { MessageActionRow, MessageEmbed, MessageButton } = require('discord.js');
+const { MessageEmbed }  = require('discord.js');
+const generatePages = require('../utils/embedPages.js');
 
 module.exports = new Command({
 	name: "queue",
     aliases: ['q'],
 	description: "Displays the server queue",
 	permission: "SEND_MESSAGES",
-	async run(message, args, client, _fromButton = false) {
+	async run(message, args, client, slash, _fromButton = false) {
         const queue = client.player.getQueue(message.guild);
         if (!queue || !queue.current) {
             if(_fromButton) return;
@@ -14,29 +15,9 @@ module.exports = new Command({
             embed.setTitle('Server Queue');
             embed.setColor('#b84e44');
             embed.setDescription(`No songs in the queue.`);
-            return message.channel.send({ embeds: [embed] });
+            return message.reply({ embeds: [embed] });
         }
         let usedby = _fromButton ? message.member : ""; // should test first
-
-        const buttons = [
-            new MessageButton()
-                .setCustomId('first')
-                .setLabel('<<')
-                .setStyle('SECONDARY'),
-            new MessageButton()
-                .setCustomId('previous')
-                .setLabel('<')
-                .setStyle('SECONDARY'),
-            new MessageButton()
-                .setCustomId('next')
-                .setLabel('>')
-                .setStyle('SECONDARY'),
-            new MessageButton()
-                .setCustomId('last')
-                .setLabel('>>')
-                .setStyle('SECONDARY')
-        ];
-        const row = new MessageActionRow().addComponents(buttons);
 
         const pages = [];
         let page = 1;
@@ -67,65 +48,14 @@ module.exports = new Command({
                     embed.setColor('#44b868');
                     embed.setDescription(`${usedby}\nNo more songs in the queue.`);
                     embed.setAuthor({ name: `Now playing: ${queue.current.title}`, iconURL: null, url: `${queue.current.url}` });
-                    return message.channel.send({ embeds: [embed] });
+                    return message.reply({ embeds: [embed] });
                 }
                 if(page === 2) {
-                    return message.channel.send({ embeds: [pages[0]] });
+                    return message.reply({ embeds: [pages[0]] });
                 }
             }
         } while(!emptypage);
 
-        let currentPage = 0;
-        const queueList = await message.channel.send({
-            embeds: [pages[currentPage].setFooter({ text: `Page ${currentPage+1}/${pages.length}` })],
-            components: [row]
-        });
-        const filter = (button) => button.customId === 'first' || 'previous' || 'next' || 'last';
-        const collector = await queueList.createMessageComponentCollector({ filter, time: 40000 });
-
-        collector.on("collect", async (button) => {
-            switch(button.customId) {
-                case 'first':
-                    currentPage = 0;
-                    break;
-                case 'previous':
-                    currentPage = currentPage > 0 ? --currentPage : pages.length-1;
-                    break;
-                case 'next':
-                    currentPage = currentPage+1 < pages.length ? ++currentPage : 0;
-                    break;
-                case 'last':
-                    currentPage = pages.length-1;
-                    break;
-            }
-            switch(currentPage) {
-                case 0:
-                    row.setComponents(buttons[0].setDisabled(true), buttons[1].setDisabled(true), buttons[2].setDisabled(false), buttons[3].setDisabled(false));
-                    break;
-                case pages.length-1:
-                    row.setComponents(buttons[0].setDisabled(false), buttons[1].setDisabled(false), buttons[2].setDisabled(true), buttons[3].setDisabled(true));
-                    break;
-                default:
-                    row.setComponents(buttons[0].setDisabled(false), buttons[1].setDisabled(false), buttons[2].setDisabled(false), buttons[3].setDisabled(false));
-                    break;
-            }
-            queueList.edit({
-                embeds: [pages[currentPage].setFooter({ text: `Page ${currentPage+1}/${pages.length}` })],
-                components: [row]
-            });
-            collector.resetTimer();
-            await button.deferUpdate();
-        });
-
-        collector.on("end", (_, reason) => {
-            if(reason === "time") {
-                row.setComponents(buttons[0].setDisabled(true), buttons[1].setDisabled(true), buttons[2].setDisabled(true), buttons[3].setDisabled(true));
-                queueList.edit({
-                    embeds: [pages[currentPage].setFooter({ text: `Page ${currentPage+1}/${pages.length}` })],
-                    components: [row]
-                }).catch(error=> {});
-            }
-        });
-
+        generatePages(message, pages, { timeout: 40000, fromButton: _fromButton } );
 	}
 });
